refactor database schema
Per convo with Laura, Zac and Matt, decided to have the database only consist of the merged detectiona and deployment tables, and then have the tags and deployment tables included as .csvs.  This is because the merged table is what's useful in their analyses, and it's too large for a .csv.