Fixes two receiver locations and updates file paths
With the update in the raw data and re-export of all_rec_locs.csv, two incorrect receiver locations are fixed (closes #8 and closes #9).

Also updated filepaths to the raw data in clean_deployments, because we are hosting all raw data in the dropbox folder so that R/set_data_dir.R works as intended with multiple contributors.

Also moved some optional bounding code from clean_deployments.R to receiver_map.R in order to keep all mapping code in the same script; it is wrapped in if(FALSE) because it's not currently used.